// frontend/src/types/pbac.types.ts
export interface IPermission {
  code: string
  name: string
  description: string
}